Fix FTPSensor failing on error message with unexpected text. (#2450)
* [AIRFLOW-1413] Fix FTPSensor file presence check

Currently FTPSensor operates by checking text of error
message returned from ftp lib. It only succeeds if the
message matches the expected text. Otherwise it fails
with an exception. However the message is dependend
on a system, locale and possibly other factors.

This patch changes the operation to inspect error code
rather than message text.

It also adds option to ignore certain classes of errors
such as Host Unavailable that are recoverable, thus the
performed action can and should be retried according
to ftp spec.

* [AIRFLOW-1413] Adjustments as per code review

* [AIRFLOW-1413] fixing style

Co-Authored-By: mdziemianko <michal.dziemianko@gmail.com>